requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
et_xmlfile==2.0.0
idna==3.10
joblib==1.5.2
nltk==3.9.2
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
textblob==0.19.0
textstat==0.7.10
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
newspaper3k
lxml_html_clean
